BUILD: clean-up cmake structure and improve (-DPREFIX+... support for OSXFuse etc...) PROTABILITY: add tons' of exceptions to compile on OSX 10.8

+++ common/dbmaptest/DbMapTestBurn.cc
@@ -76,7 +76,7 @@ void* TestWriteGlob(void *threadid) {
     char vs[]="value_xxxxxx";
     char cs[]="comment_xxxxxx";
 
-    unsigned long int thrid=pthread_self();
+    unsigned long int thrid= (unsigned long int) pthread_self();
 
 	for(int k=0;k<total;k++) {
 	    AtomicInc(writecount);
